Update zpl platforms
Things are shaping up fast for the Posix layer.  For sanity, I have
added fstest as a simple test to verify common operations are working
correctly in the Posix layer.  It runs roughly 1900 simple tests and
only takes a few minutes.  Of the 5 distributions I've been working
with currently all 5 pass the entire suite.