Remove the init-top/hostname script and force BUSYBOX=y instead.
The pre-init script in the initramfs environment sets the hostname
if the regular /bin/hostname binary and the /etc/hostname file are
included by the hook.

The pre-init invokes `/bin/hostname -b -F /etc/hostname`, which is
incompatible with the busybox builtin because it lacks the '-b'
switch.

This commit mysteriously resolves the console corruption that
occasionally happens on systems that boot into a native ZFS root
filesystem.