Update debian/changelog using git-dch.
Correct typos in the debian/debuild-ppa.sh helper.

Set extend-diff-ignore='.*' in the debian/source/options file to
disable the automatic patch generated by the implied
--single-debian-patch during PPA builds for more than one series.

The debian/patches/debian-changes-* files do not contain a useful
history and only add cruft to the PPA builds.  If this source
package is added to a regular repository, then the
extend-diff-ignore option should be disabled.