Add Lustre documentation
For RHEL6 there are now tagged versions of the spl, zfs, lustre-ldisk,
ldiskfsprogs, and lustre-kmu.  These have been tested to be largely
functional under RHEL6.  This commit adds some basic documentation to
the website which refers people to correct tags to use, how to build
each package, and provides a simple example configuration.

The Lustre-kdmu work is still under serious development and many
sharp corners remain.  However, for the first time all of this work
has been moved out fully in to the open and basic documentation
provided.  Any sites which are interested in this work are welcome
to try code and contribute.