Fix `make distclean` for `./configure --with-config=user`.
Running `make distclean` fails after `./configure --with-config=user`
with this error message:

  Making distclean in module
  make[1]: Entering directory `/zfs/module'
  make -C  SUBDIRS=`pwd`  clean
  make: Entering an unknown directory
  make: *** SUBDIRS=/zfs/module: No such file or directory.  Stop.
  make: Leaving an unknown directory
  make[1]: *** [clean] Error 2
  make[1]: Leaving directory `/zfs/module'
  make: *** [distclean-recursive] Error 1

This happens because the DIST_SUBDIRS variable is used instead of
the regular SUBDIRS variable in the $(RECURSIVE_CLEAN_TARGETS) rule.

SUBDIRS is pruned according to configuration parameters, but
DIST_SUBDIRS is not because it contains a list of all components
for the distdir rules.

Correcting this is worthwhile because debhelper assumes that all
configurations produce a working distclean rule, and the debian/
packaging overlay currently needs a workaround for it.