Move all development items to the issue tracker
I've moved all the pending development items to the issue tracker.
This ensures everything is in one place and makes it easy for
people to see what still needs to be done and comment on the
issues.  In the best case someone may even volunteer to help.
All non-critical issues are marked as features.