Add a ZPL example
Update the example zpl documentation with a trivial example of how
to create a pool and filesystem.  Better documentation is available
elsewhere but this at least is enough to get your a filesystem.

Also removed the KQ and zfs-fuse references which are increasingly
out of place on this page.  Drop the graphic as well, I'm moving it
to the example zpios page which it's more helpful.